// web-react/src/rpc/client.gen.ts
export interface SongListItem {
  slug: string
  title: string
  artist: string
  key?: string
}